move validation out of schemas
